Use poll with nfds_t type
Clang 16 makes -Wincompatible-function-pointer-types an error by
default, which causes build failure for `(*_poll)` function because
`nfds_t` does not need to be equivalent to `unsigned int`. This commit
changes it in all poll function definitions and prototypes to `nfds_t`
as it is defined by POSIX and defines `nfds_t` type as `unsigned int`
when `WITH_POLL` is not defined.

Fixes: https://github.com/3proxy/3proxy/issues/895
See-also: https://bugs.gentoo.org/881015